+++ sklearn/tree/__init__.py
@@ -3,11 +3,11 @@ The :mod:`sklearn.tree` module includes decision tree-based models for
 classification and regression.
 """
 
-from ._base import BaseDecisionTree
-from ._base import DecisionTreeClassifier
-from ._base import DecisionTreeRegressor
-from ._base import ExtraTreeClassifier
-from ._base import ExtraTreeRegressor
+from ._classes import BaseDecisionTree
+from ._classes import DecisionTreeClassifier
+from ._classes import DecisionTreeRegressor
+from ._classes import ExtraTreeClassifier
+from ._classes import ExtraTreeRegressor
 from ._export import export_graphviz, plot_tree, export_text
 
 __all__ = ["BaseDecisionTree",
